chore: test perf
