+++ package.json
@@ -344,7 +344,8 @@
   },
   "scripts": {
     "build": "npx tsc -noEmit && node build.mjs",
-    "test": "node --no-experimental-strip-types ./node_modules/jasmine/bin/jasmine.js",
+    "test": "node test.cjs",
+    "test:seq": "node --no-experimental-strip-types ./node_modules/jasmine/bin/jasmine.js",
     "lint": "eslint src",
     "lint:errors": "eslint --quiet src",
     "vscode:prepublish": "yarn build",
